fix directfb input

+++ src/SafeInt.h
@@ -727,7 +727,7 @@ template <> class MinMax< true,  16 > { public: const static __int16 min = ( -0x
                                                 const static __int16 max = 0x7fff; };
 template <> class MinMax< true,  32 > { public: const static __int32 min = ( -0x7fffffff -1 );
                                                 const static __int32 max = 0x7fffffff; };
-template <> class MinMax< true,  64 > { public: const static __int64 min = 0x8000000000000000LL;
+template <> class MinMax< true,  64 > { public: const static __int64 min = (const __int64) 0x8000000000000000LL;
                                                 const static __int64 max = 0x7fffffffffffffffLL; };
 
 template <> class MinMax< false,  8 > { public: const static unsigned __int8  min = 0;
